refactor(live): extract validation contract layer
