+++ scripts/live/validate-canonical-workbench-live.mjs
@@ -1,334 +1,43 @@
 import dns from "node:dns/promises";
-import fs from "node:fs/promises";
 import path from "node:path";
 import process from "node:process";
 import { chromium, expect } from "@playwright/test";
+import { resolveValidationConfig } from "./validation/args.mjs";
+import {
+  loadCanonicalContractMetadata,
+  loadWorkbenchPanelRegistryMetadata,
+} from "./validation/contract-metadata.mjs";
+import {
+  buildSummaryPaths,
+  createValidationSummary,
+  ensureDirectory,
+  writeShotIndex,
+  writeValidationSummary,
+} from "./validation/evidence-summary-writer.mjs";
 
-const DEFAULT_CANONICAL_CONTRACT = {
-  contractId: "canonical-front-office-demo-data-contract",
-  contractVersion: "1.0.0",
-  governedByRfc: "RFC-0076",
-  portfolioId: "PB_SG_GLOBAL_BAL_001",
-  benchmarkCode: "BMK_PB_GLOBAL_BALANCED_60_40",
-  canonicalAsOfDate: "2026-04-10",
-};
-
-const DEFAULT_PANEL_REGISTRY = {
-  contractId: "workbench-panel-registry",
-  contractVersion: "1.0.0",
-  governedByRfc: "RFC-0077",
-  canonicalDataContract: "canonical-front-office-demo-data-contract",
-  sourcePath: "deterministic-fallback",
-  panels: [
-    {
-      panelId: "portfolio.summary",
-      owningService: "lotus-gateway",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/overview",
-      requiredSupportState: "ready",
-      route: "/portfolio?portfolioId={portfolioId}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "portfolio-summary-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "portfolio.detailed",
-      owningService: "lotus-gateway",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/overview",
-      requiredSupportState: "ready",
-      route: "/portfolio?portfolioId={portfolioId}&tab=detailed",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "portfolio-detailed-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.summary",
-      owningService: "lotus-performance",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/performance/summary",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-summary-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.analysis.contribution",
-      owningService: "lotus-performance",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/performance/details",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=analysis&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-analysis-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.analysis.attribution",
-      owningService: "lotus-performance",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/performance/details",
-      requiredSupportState: "partial",
-      route: "/performance?portfolioId={portfolioId}&mode=analysis&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-analysis-live.png",
-      knownLimitations: [
-        "benchmark-relative attribution may remain partial until full source-backed detail is available",
-      ],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.advisor_brief",
-      owningService: "lotus-performance",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/performance/advisor-brief",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=advisor&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-advisor-brief-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.risk.snapshot",
-      owningService: "lotus-risk",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/risk/summary",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=risk&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-risk-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.risk.drawdown",
-      owningService: "lotus-risk",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/risk/drawdown",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=risk&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-risk-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.risk.concentration",
-      owningService: "lotus-risk",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/risk/concentration",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=risk&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-risk-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.risk.rolling",
-      owningService: "lotus-risk",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/risk/rolling",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=risk&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-risk-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.risk.historical_attribution",
-      owningService: "lotus-risk",
-      gatewayEndpoint: "/api/v1/workbench/{portfolio_id}/risk/attribution",
-      requiredSupportState: "ready",
-      route: "/performance?portfolioId={portfolioId}&mode=risk&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-risk-live.png",
-      knownLimitations: [],
-      ownerFollowUpRfc: null,
-    },
-    {
-      panelId: "performance.evidence",
-      owningService: "lotus-gateway",
-      gatewayEndpoint: null,
-      requiredSupportState: "unavailable",
-      route: "/performance?portfolioId={portfolioId}&mode=evidence&period=YTD&detailBasis=NET&benchmark={benchmarkCode}",
-      allowedStates: ["ready", "loading", "empty", "partial", "unavailable", "error"],
-      screenshotName: "performance-evidence-live.png",
-      knownLimitations: ["full evidence and lineage support is deferred pending RFC-0079"],
-      ownerFollowUpRfc: "RFC-0079",
-    },
-  ],
-};
-
-function parseArgs(argv) {
-  const args = new Map();
-  for (let index = 0; index < argv.length; index += 1) {
-    const token = argv[index];
-    if (!token.startsWith("--")) continue;
-    const key = token.slice(2);
-    const next = argv[index + 1];
-    if (!next || next.startsWith("--")) {
-      args.set(key, "true");
-      continue;
-    }
-    args.set(key, next);
-    index += 1;
-  }
-  return args;
-}
-
-const args = parseArgs(process.argv.slice(2));
-const portfolioId = args.get("portfolio-id") ?? "PB_SG_GLOBAL_BAL_001";
-const benchmarkCode = args.get("benchmark-code") ?? "BMK_PB_GLOBAL_BALANCED_60_40";
-const workbenchBaseUrl = (args.get("workbench-base-url") ?? "http://workbench.dev.lotus").replace(/\/+$/, "");
-const gatewayBaseUrl = (args.get("gateway-base-url") ?? "http://gateway.dev.lotus").replace(/\/+$/, "");
-const outputDir = path.resolve(
-  process.cwd(),
-  args.get("output-dir") ?? "output/playwright/live-canonical"
-);
-const summaryPath = path.join(outputDir, "live-validation-summary.json");
-const shotIndexPath = path.join(outputDir, "SHOT-INDEX.md");
-const timeoutMs = Number(args.get("timeout-ms") ?? "60000");
-const canonicalAsOfDate = args.get("as-of-date") ?? "2026-04-10";
+const {
+  portfolioId,
+  benchmarkCode,
+  workbenchBaseUrl,
+  gatewayBaseUrl,
+  outputDir,
+  timeoutMs,
+  canonicalAsOfDate,
+} = resolveValidationConfig(process.argv.slice(2));
+const { summaryPath, shotIndexPath } = buildSummaryPaths(outputDir);
 const canonicalContract = await loadCanonicalContractMetadata();
 const panelRegistry = await loadWorkbenchPanelRegistryMetadata();
 const panelRegistryById = new Map(panelRegistry.panels.map((panel) => [panel.panelId, panel]));
 
-const summary = {
+const summary = createValidationSummary({
   generatedAt: new Date().toISOString(),
   portfolioId,
   benchmarkCode,
   canonicalContract,
-  panelRegistry: {
-    contractId: panelRegistry.contractId,
-    contractVersion: panelRegistry.contractVersion,
-    governedByRfc: panelRegistry.governedByRfc,
-    canonicalDataContract: panelRegistry.canonicalDataContract,
-    sourcePath: panelRegistry.sourcePath,
-  },
   workbenchBaseUrl,
   gatewayBaseUrl,
-  dns: [],
-  apiChecks: [],
-  uiChecks: [],
-  calculationChecks: [],
-  panelClassifications: [],
-  supportabilityChecks: [],
-  screenshots: [],
-};
-
-async function loadCanonicalContractMetadata() {
-  const candidatePaths = [
-    process.env.LOTUS_PLATFORM_REPO
-      ? path.resolve(
-          process.env.LOTUS_PLATFORM_REPO,
-          "context",
-          "contracts",
-          "canonical-front-office-demo-data-contract.json"
-        )
-      : null,
-    path.resolve(
-      process.cwd(),
-      "..",
-      "lotus-platform",
-      "context",
-      "contracts",
-      "canonical-front-office-demo-data-contract.json"
-    ),
-  ].filter(Boolean);
-
-  for (const candidatePath of candidatePaths) {
-    try {
-      const raw = await fs.readFile(candidatePath, "utf8");
-      const payload = JSON.parse(raw);
-      return {
-        contractId: payload.contract_id ?? DEFAULT_CANONICAL_CONTRACT.contractId,
-        contractVersion: payload.contract_version ?? DEFAULT_CANONICAL_CONTRACT.contractVersion,
-        governedByRfc: payload.governed_by_rfc ?? DEFAULT_CANONICAL_CONTRACT.governedByRfc,
-        portfolioId: payload.portfolio?.portfolio_id ?? DEFAULT_CANONICAL_CONTRACT.portfolioId,
-        benchmarkCode:
-          payload.benchmark?.benchmark_id ?? DEFAULT_CANONICAL_CONTRACT.benchmarkCode,
-        canonicalAsOfDate:
-          payload.date_policy?.canonical_as_of_date ??
-          DEFAULT_CANONICAL_CONTRACT.canonicalAsOfDate,
-        sourcePath: candidatePath,
-      };
-    } catch (error) {
-      if (error && typeof error === "object" && "code" in error && error.code === "ENOENT") {
-        continue;
-      }
-      throw new Error(
-        `Unable to load governed canonical contract metadata from ${candidatePath}: ${
-          error instanceof Error ? error.message : String(error)
-        }`
-      );
-    }
-  }
-
-  return {
-    ...DEFAULT_CANONICAL_CONTRACT,
-    sourcePath: "deterministic-fallback",
-  };
-}
-
-async function loadWorkbenchPanelRegistryMetadata() {
-  const candidatePaths = [
-    process.env.LOTUS_PLATFORM_REPO
-      ? path.resolve(
-          process.env.LOTUS_PLATFORM_REPO,
-          "context",
-          "contracts",
-          "workbench-panel-registry.json"
-        )
-      : null,
-    path.resolve(
-      process.cwd(),
-      "..",
-      "lotus-platform",
-      "context",
-      "contracts",
-      "workbench-panel-registry.json"
-    ),
-  ].filter(Boolean);
-
-  for (const candidatePath of candidatePaths) {
-    try {
-      const raw = await fs.readFile(candidatePath, "utf8");
-      const payload = JSON.parse(raw);
-      return {
-        contractId: payload.contract_id ?? DEFAULT_PANEL_REGISTRY.contractId,
-        contractVersion: payload.contract_version ?? DEFAULT_PANEL_REGISTRY.contractVersion,
-        governedByRfc: payload.governed_by_rfc ?? DEFAULT_PANEL_REGISTRY.governedByRfc,
-        canonicalDataContract:
-          payload.canonical_data_contract ?? DEFAULT_PANEL_REGISTRY.canonicalDataContract,
-        sourcePath: candidatePath,
-        panels: (payload.panels ?? []).map((panel) => ({
-          panelId: panel.panel_id,
-          owningService: panel.owning_service,
-          gatewayEndpoint: panel.gateway_endpoint,
-          requiredSupportState: panel.required_support_state,
-          route: panel.route,
-          allowedStates: panel.allowed_states ?? [],
-          screenshotName: panel.screenshot_policy?.screenshot_name ?? null,
-          knownLimitations: panel.known_limitations ?? [],
-          ownerFollowUpRfc: panel.owner_follow_up_rfc ?? null,
-        })),
-      };
-    } catch (error) {
-      if (error && typeof error === "object" && "code" in error && error.code === "ENOENT") {
-        continue;
-      }
-      throw new Error(
-        `Unable to load governed panel registry metadata from ${candidatePath}: ${
-          error instanceof Error ? error.message : String(error)
-        }`
-      );
-    }
-  }
-
-  return DEFAULT_PANEL_REGISTRY;
-}
-
-async function ensureDirectory(target) {
-  await fs.mkdir(target, { recursive: true });
-}
+  panelRegistry,
+});
 
 async function checkDns(hostname, required = true) {
   try {
@@ -773,35 +482,6 @@ async function screenshotRegisteredPanel(page, panelId, metadata = {}) {
   });
 }
 
-async function writeSummary() {
-  await fs.writeFile(summaryPath, `${JSON.stringify(summary, null, 2)}\n`, "utf8");
-}
-
-async function writeShotIndex() {
-  const lines = [
-    "# Lotus Canonical Front-Office Screenshots",
-    "",
-    `- Generated: ${summary.generatedAt}`,
-    `- Contract: ${summary.canonicalContract.contractId} ${summary.canonicalContract.contractVersion}`,
-    `- Governed by: ${summary.canonicalContract.governedByRfc}`,
-    `- Portfolio: ${portfolioId}`,
-    `- Benchmark: ${benchmarkCode}`,
-    `- As of: ${canonicalAsOfDate}`,
-    `- Validation summary: ${summaryPath}`,
-    "",
-    "## Captures",
-    "",
-  ];
-
-  for (const screenshotEvidence of summary.screenshots) {
-    lines.push(
-      `- ${screenshotEvidence.name} - ${screenshotEvidence.panel} - ${screenshotEvidence.route} - ${screenshotEvidence.state}`
-    );
-  }
-
-  await fs.writeFile(shotIndexPath, `${lines.join("\n")}\n`, "utf8");
-}
-
 async function run() {
   await ensureDirectory(outputDir);
 
@@ -1131,8 +811,8 @@ async function run() {
     await browser.close();
   }
 
-  await writeSummary();
-  await writeShotIndex();
+  await writeValidationSummary(summaryPath, summary);
+  await writeShotIndex(shotIndexPath, summary, summaryPath);
   console.log(`Live canonical Workbench validation passed for ${portfolioId}. Screenshots: ${outputDir}`);
 }
