+++ scripts/live/validation/evidence-summary-writer.mjs
@@ -1,0 +1,75 @@
+import fs from "node:fs/promises";
+import path from "node:path";
+
+export function createValidationSummary({
+  generatedAt = new Date().toISOString(),
+  portfolioId,
+  benchmarkCode,
+  canonicalContract,
+  panelRegistry,
+  workbenchBaseUrl,
+  gatewayBaseUrl,
+}) {
+  return {
+    generatedAt,
+    portfolioId,
+    benchmarkCode,
+    canonicalContract,
+    panelRegistry: {
+      contractId: panelRegistry.contractId,
+      contractVersion: panelRegistry.contractVersion,
+      governedByRfc: panelRegistry.governedByRfc,
+      canonicalDataContract: panelRegistry.canonicalDataContract,
+      sourcePath: panelRegistry.sourcePath,
+    },
+    workbenchBaseUrl,
+    gatewayBaseUrl,
+    dns: [],
+    apiChecks: [],
+    uiChecks: [],
+    calculationChecks: [],
+    panelClassifications: [],
+    supportabilityChecks: [],
+    screenshots: [],
+  };
+}
+
+export async function ensureDirectory(target) {
+  await fs.mkdir(target, { recursive: true });
+}
+
+export function buildSummaryPaths(outputDir) {
+  return {
+    summaryPath: path.join(outputDir, "live-validation-summary.json"),
+    shotIndexPath: path.join(outputDir, "SHOT-INDEX.md"),
+  };
+}
+
+export async function writeValidationSummary(summaryPath, summary) {
+  await fs.writeFile(summaryPath, `${JSON.stringify(summary, null, 2)}\n`, "utf8");
+}
+
+export async function writeShotIndex(shotIndexPath, summary, validationSummaryPath) {
+  const lines = [
+    "# Lotus Canonical Front-Office Screenshots",
+    "",
+    `- Generated: ${summary.generatedAt}`,
+    `- Contract: ${summary.canonicalContract.contractId} ${summary.canonicalContract.contractVersion}`,
+    `- Governed by: ${summary.canonicalContract.governedByRfc}`,
+    `- Portfolio: ${summary.portfolioId}`,
+    `- Benchmark: ${summary.benchmarkCode}`,
+    `- As of: ${summary.screenshots[0]?.asOfDate ?? summary.canonicalContract.canonicalAsOfDate ?? "unknown"}`,
+    `- Validation summary: ${validationSummaryPath}`,
+    "",
+    "## Captures",
+    "",
+  ];
+
+  for (const screenshotEvidence of summary.screenshots) {
+    lines.push(
+      `- ${screenshotEvidence.name} - ${screenshotEvidence.panel} - ${screenshotEvidence.route} - ${screenshotEvidence.state}`
+    );
+  }
+
+  await fs.writeFile(shotIndexPath, `${lines.join("\n")}\n`, "utf8");
+}
